# src/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/create_user":{"post":{"summary":"Create User","operationId":"create_user_api_create_user_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/read_users":{"get":{"summary":"Read Users","operationId":"read_users_api_read_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Users"},"type":"array","title":"Response Read Users Api Read Users Get"}}}}}}},"/api/update_user":{"put":{"summary":"Update User","operationId":"update_user_api_update_user_put","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/delete_user":{"delete":{"summary":"Delete User","operationId":"delete_user_api_delete_user_delete","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/create_lottery":{"post":{"summary":"Create Lottery","operationId":"create_lottery_api_create_lottery_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LotteryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LotteryCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/read_lotteries":{"get":{"summary":"Read Lotteries","operationId":"read_lotteries_api_read_lotteries_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Lotteries"},"type":"array","title":"Response Read Lotteries Api Read Lotteries Get"}}}}}}},"/api/read_lotteries_by_user_id":{"get":{"summary":"Read Lotteries By User Id","operationId":"read_lotteries_by_user_id_api_read_lotteries_by_user_id_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Lotteries"},"title":"Response Read Lotteries By User Id Api Read Lotteries By User Id Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/update_lottery":{"put":{"summary":"Update Lottery","operationId":"update_lottery_api_update_lottery_put","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LotteryCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LotteryCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/delete_lottery":{"delete":{"summary":"Delete Lottery","operationId":"delete_lottery_api_delete_lottery_delete","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Lotteries":{"properties":{"user_id":{"type":"integer","title":"User Id","default":-1,"desciption":"このくじ引きの所有者ID"},"text":{"type":"string","title":"Text","default":"","desciption":"くじ引きの抽選対象"},"title":{"type":"string","title":"Title","default":"","desciption":"くじ引き名"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"Lotteries"},"LotteryCreate":{"properties":{"user_id":{"type":"integer","title":"User Id","default":-1,"desciption":"このくじ引きの所有者ID"},"text":{"type":"string","title":"Text","default":"","desciption":"くじ引きの抽選対象"},"title":{"type":"string","title":"Title","default":"","desciption":"くじ引き名"}},"type":"object","title":"LotteryCreate"},"LotteryCreateResponse":{"properties":{"user_id":{"type":"integer","title":"User Id","default":-1,"desciption":"このくじ引きの所有者ID"},"text":{"type":"string","title":"Text","default":"","desciption":"くじ引きの抽選対象"},"title":{"type":"string","title":"Title","default":"","desciption":"くじ引き名"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"LotteryCreateResponse"},"UserCreate":{"properties":{"account_name":{"type":"string","title":"Account Name","default":"","desciption":"アカウント名"},"identification":{"type":"string","title":"Identification","default":"","desciption":"識別情報"}},"type":"object","title":"UserCreate"},"UserCreateResponse":{"properties":{"account_name":{"type":"string","title":"Account Name","default":"","desciption":"アカウント名"},"identification":{"type":"string","title":"Identification","default":"","desciption":"識別情報"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"UserCreateResponse"},"Users":{"properties":{"account_name":{"type":"string","title":"Account Name","default":"","desciption":"アカウント名"},"identification":{"type":"string","title":"Identification","default":"","desciption":"識別情報"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"Users"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}